requirements.txt
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.2.1
Django==2.2
django-crispy-forms==1.9.2
django-crontab==0.7.1
django-extensions==3.0.9
django-mathfilters==1.0.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-weekday-field==1.1.0
djangorestframework==3.12.2
html5lib==1.1
idna==2.10
mysqlclient==2.0.3
paytmchecksum==1.7.0
pickleshare==0.7.5
Pillow==7.2.0
pycparser==2.20
pycryptodome==3.9.8
PyPDF2==1.26.0
pytz==2020.5
reportlab==3.5.59
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
webencodings==0.5.1
xhtml2pdf==0.2.4
